+++ examples/Vendored XCFramework Example/BananaLib/StaticLibrary/CoconutLib.xcframework/watchos-i386-simulator/Headers/CoconutLib.h
@@ -1,9 +1,0 @@
-#import <Foundation/Foundation.h>
-
-//! Project version number for CoconutLib.
-FOUNDATION_EXPORT double CoconutLibVersionNumber;
-
-//! Project version string for CoconutLib.
-FOUNDATION_EXPORT const unsigned char CoconutLibVersionString[];
-
-#import <CoconutLib/Coconut.h>